requirements.txt
streamlit
google-generativeai
chromadb
PyPDF2
